+++ lib/src/filters/production_filter.dart
@@ -1,3 +1,10 @@
+/*
+ * @Author: Cao Shixin
+ * @Date: 2021-01-18 16:27:06
+ * @LastEditors: Cao Shixin
+ * @LastEditTime: 2021-03-25 17:55:47
+ * @Description: 
+ */
 import 'package:logger_csx/src/logger.dart';
 import 'package:logger_csx/src/log_filter.dart';
 
@@ -5,6 +12,6 @@ import 'package:logger_csx/src/log_filter.dart';
 class ProductionFilter extends LogFilter {
   @override
   bool shouldLog(LogEvent event) {
-    return event.level.index >= level.index;
+    return event.level.index >= level!.index;
   }
 }