+++ lib/src/log_filter.dart
@@ -5,7 +5,7 @@ import 'package:logger_csx/src/logger.dart';
 /// You can implement your own `LogFilter` or use [DevelopmentFilter].
 /// Every implementation should consider [Logger.level].
 abstract class LogFilter {
-  late Level level;
+  Level? level;
   void init() {}
 
   /// Is called every time a new log message is sent and decides if